improved separation of concerns
anyway the system does not yet detect, when to send an email

+++ clear-statestores.sh
@@ -3,4 +3,6 @@
 docker exec -it dapr_redis redis-cli FLUSHALL
 # sqlite3 .data/dapr_statestore.db "DELETE FROM state;"
 # rm -rf .data/dapr_statestore.db
-rm *_state.json
+if ls *_state.json 1> /dev/null 2>&1; then
+	rm *_state.json
+fi
